updated the eventDetailsModal and EventsTable component
Page.js: Added emissions as a prop for EventsTable.

carbonEmissionsModal.js: Changed the label description

eventDetailsModal.js: added emissionsData prop to take in potential emissions data. Added method to dynamically render section to display emission data in the modal if there was any emissions data.

table.js: Mapped emissions data into a store.
Added state for emissionData when a row is selected
updated function handleRowClick to check if key from emissionsData mapping is the same as the key from an event. If so that emissions data would be passed as selected Emissions data